adding exception handlers and fixes, adding validation

+++ webapi/src/main/java/com/WebApi/webapi/services/ProductService.java
@@ -1,12 +1,15 @@
 package com.WebApi.webapi.services;
 
+import com.WebApi.webapi.exceptionsHandlers.ServiceException;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import com.WebApi.webapi.models.Product;
 import com.WebApi.webapi.models.ProductValidator;
-import org.springframework.beans.factory.annotation.Autowired;
+
 import org.springframework.stereotype.Service;
 import com.WebApi.webapi.repositories.ProductRepository;
+
+
 
 import java.util.List;
 
@@ -15,17 +18,27 @@ import java.util.List;
 @Slf4j
 public class ProductService {
 
-    @Autowired
     private final ProductRepository productRepository;
 
-    public List<Product> getAllProducts() {
+
+    public List<Product> getAllProducts() throws Exception {
         log.info("getting all products | ProductService");
-        return productRepository.findAll();
+        List<Product> lst = productRepository.findAll();
+        if(lst.size()==0){
+            throw new ServiceException("None elements found");
+        }
+        return lst;
     }
 
-    public Product getProductById(long id){
+    public Product getProductById(long id) throws ServiceException {
         log.info("getting product by id:"+id+" | ProductService");
-        return productRepository.findById(id);
+        if(productRepository.findById(id)!=null){
+            return productRepository.findById(id);
+        }
+        else{
+            throw new ServiceException("Product not found, given id:"+id);
+        }
+
     }
 
     public Long addNewProduct(ProductValidator model){
@@ -37,16 +50,28 @@ public class ProductService {
 
     }
 
-    public void updateProduct(long id, ProductValidator model) {
+    public void updateProduct(long id, ProductValidator model) throws Exception {
         log.info("updating product by id:"+id+" | ProductService");
         Product product = productRepository.findById(id);
-        product.setPrice(model.getM_price());
-        product.setName(model.getM_name());
-        productRepository.saveAndFlush(product);
+        if(product != null){
+            product.setPrice(model.getM_price());
+            product.setName(model.getM_name());
+            productRepository.saveAndFlush(product);
+        }else{
+            throw new ServiceException("Product not found, given id:"+id);
+        }
     }
 
-    public void deleteProduct(long id) {
+    public void deleteProduct(long id) throws Throwable {
         log.info("deleting product by id:"+id+" | ProductService");
-        productRepository.deleteById(id);
+
+            if(productRepository.findById(id)!=null){
+                productRepository.deleteById(id);
+            }
+            else{
+                throw new ServiceException("Product not found, given id:"+id);
+            }
+
+
     }
 }
